Adding typescript to the project.
