+++ package.json
@@ -12,6 +12,7 @@
     "@types/node": "^17.0.23",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
+    "@types/react-router": "^5.1.18",
     "bootstrap": "^4.1.1",
     "font-awesome": "^4.7.0",
     "lodash": "^4.17.10",
@@ -53,6 +54,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "gh-pages": "^3.1.0",
     "parcel": "^1.12.3"
   }
