+++ src/App.jsx
@@ -1,23 +1,19 @@
-import { getMovies } from './services/fakeMovieService.js';
-import {useState} from 'react';
 import TableName from './components/tableName';
 import Movies from './components/movies';
 import Todo from './components/todoappUI/todo';
-import ContactMe from './components/contactMe.jsx';
+import ContactMe from './components/contactMe.tsx';
 import Profile from './components/Profile/profile';
 import MainHeader from './components/header/mainheader';
 import SideBar from './components/sidebar/sideBar.jsx';
-import React, { Component } from 'react';
+import React from 'react';
 import './App.scss';
 
 import {
   Route,
-  NavLink,
   HashRouter
 } from "react-router-dom";
 
-class App extends Component {
-  render(){
+export default function App() {
     return (
         <HashRouter>
           <main className="App">
@@ -32,16 +28,13 @@ class App extends Component {
                 <Route exact path='/' component={Profile}></Route>
                 <Route path='/todo' component={Todo}></Route>
                 <Route path='/movies' component={Movies}></Route>
-                <Route path='/contactme' component={() => <ContactMe text="CONTACT_ME"/>}></Route>
+                <Route path='/contactme' component={() => <ContactMe/>}></Route>
                 <Route path='/tablename' component={TableName}></Route>
             </div>
             </section>
           </section>
         </HashRouter>
     );
-  }
 }
-
-export default App;
